Add import default_these module
Looking at ways to reduce the size of some of the code by pulling some things out into modules.  This way perhaps some expandability can be added to the program.  Certainly there is a concern that if it is split up too much it will be unbearable to track and follow, but ideally just some big chunks can be moved out.  Further a goal is to possibly make this code accessible from areas outside of the EP4 main control window, thus action/partial action can be implemented by other means.