Use of alternate primary key
Primary key is now exported with full product export if not set to the default of products_model; however, does not prevent the primary key from being one of the other existing $filelayout options (ie. exported twice).  While hasn't been identified as a problem in the past, perhaps a "sanitizer" for export should be used to ensure that a field is exported only once and thus not complicate the import process.

Associated with issue #26