Added additional locations to export an alternate primary key
While not completely incorporated, the use of the alternate primary key (products_id) has been further incorporated; however, if that primary key is yet something else, more work is necessary as not all cases include a notifier or "filter" to apply the potential additional primary key(s).  Part of this could eventually be addressed through the proposed expanded use of the user defined field where the table is incorporated into the field designation.

This needs to be "tested"/verified as some of the changes were more "blindly" applied than purposefully applied.