Make insert ignore and add version update
Modified the insert sql to be insert ignore into... This is to help support the update process so that the existing settings do not need to be modified/removed unless a version change does that modification/removal.  Also added a field for version tracking which will/should support notification of the need to "refresh" the configuration panel.  Now need a "flag" that there was an update such that these things can work together.